+++ Thaw/Main/AppDelegate.swift
@@ -93,6 +93,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         #endif
 
         MacOSCompatibilityWarning.showIfNeeded()
+        DebugSignals.install(appState: appState)
 
         // Warn if another menu bar manager is running.
         ConflictingAppDetector.showWarningIfNeeded()